requirements.txt
livekit>=1.0
livekit-agents[openai,deepgram,cartesia,silero,turn_detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0
asyncpg~=0.29.0
sqlalchemy[asyncio]~=2.0.23
alembic~=1.13.1
psycopg2-binary~=2.9.10
psycopg~=3.2.9
psycopg-binary~=3.2.9
psycopg-pool~=3.2.6
rich~=13.7.0
